Update README.md
In this tutorial, I'am  going to list out  the process in configuring a Node server.
From the client server (the entity engaging with the server), this particular server must have a specific function, to exhibit a strong unchanging HTML page. There is no need of CSS, front-end JavaScript, database connectivity, or any additional NPM packages beyond the essential minimum (I will provide a brief insight into what NPM entails shortly). Furthermore, it is noteworthy that this server will operate solely on a local basis without any connection to the actual internet. It is important to note that this tutorial will not delve into the fundamental mechanics of server operations, client-server communication, or other intricate aspects that are crucial for more advanced Node-related projects.